+++ core/simulation/__init__.py
@@ -1,5 +1,5 @@
-from .bus import Bus
 from .passenger_group import PassengersGroup
 from .simulation import Simulation
 from .stop import Stop
 from .line import Line, LineStop
+from .bus import Bus
