+++ demo/fiction/main.py
@@ -1,9 +1,15 @@
 import sys
 import time
 
-sys.path.append('../../')
-from core.configuration import config
-from core.simulation.simulation import Simulation
+if "core" not in sys.path[0]:
+    if "\\" in sys.path[0]:
+        sys.path.insert(0, '..\\..\\core')
+    else:
+        sys.path.insert(0, '../../core')
+
+
+from configuration import config
+from simulation.simulation import Simulation
 
 
 def read_configuration():
