--- tests/cross_version/end_to_end/test_end_to_end.py
@@ -1,1 +1,0 @@
-# TODO : end to end tests config - > symulation -> output
